Title: 길이 단위 환산하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240529/길이 단위 환산하기/convert-length-units.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double a;
-    scanf("%lf", &a);
-    printf("%.1lf", a * 30.48);
+    float a;
+    scanf("%f", &a);
+    printf("%.1f", a * 30.48);
     return 0;
 }
